Continue e/f merge
Tricky file: crypto/bn/bn_nist.c has complex changes, including a change of
function signature, and changing the type of a variable to a union.  However,
with *human* analysis of the code, I (Andrew) believe the code to be exactly
equivalent to the original.  Therefore we take version 1.0.1f.