[1.0.1ef] bugfix bss_dgram and bn_nist
bss_dgram.c defaulted to 1.0.1f when CLIVER not defined. Also, I missed a patch
entirely.

bn_nist.c: Marie protested that it was too complicated to tell whether the code
was actually equivalent, and that I had no proof that it was (it was intended
to be equivalent, but just get around some VC and Clang compiler bugs).  I am
still highly confident that it is equivalent, but we are taking the safe road
and for the large functions, we are taking the naive approach of creating an if
block around the entire function body.  The exception to this is the BIGNUM
copy functions, which are short, and which are used enough that it could hurt
our performance.