Significant changes made to ktest models.
KTOV_next_object has been changed to allow ERRORs in place of
other records, and SIGNALs before other records.  On a signal
a ktest registered signal handler will be called.

Signals: may be recorded by ktest_record_signal.  Expected
to be recorded during execution of another ktest function which writes
a record (e.g. ktest_select/accept).  A single function pointer may
be registered with ktest_register_signal_handler to be executed
on playback of a signal.

Fork: ktest_fork allows either the parent or the child's behavior
to be recorded.  The other process ceases to record to the ktest file.
It is important that on playback of a parent process ktest_fork
returns the same pid as ktest_waitpid.

Several bugs in ktest_select were fixed.