Removed signal handling as a record type, and associated support for recording. Added signal field to select record, augmented ktest_record_signal to set a static variable that ktest_select checks after return from select call in record. Changed ktest_waitpid_happytuesday back to ktest_waitpid_or_error. Added dummy ktest_shutdown.
Todo: add a proper value to ktest_select’s ktest record buffer, have added 500
just to make it work for now.

Note: in sshd, a sigchild occurs while we’re waiting on a specific select.  The
handler for that signal sets a variable in sshd indicating that the child has
died.